Moved SetupRouter into seperate file that can be imported where relevant
