+++ itu-minitwit-golang/.env.test
@@ -1,3 +1,3 @@
 PORT=8080
 DB_PATH=
-ENVIRONMENT=testing
+ENVIRONMENT=test
